+++ web/app/components/base/date-and-time-pickers/index.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useMemo, useRef, useState } from 'react'
 import dayjs, { type Dayjs } from 'dayjs'
 import { RiArrowDownSLine, RiArrowUpSLine, RiCalendarLine, RiCloseCircleFill, RiTimeLine } from '@remixicon/react'
 import cn from '@/utils/classnames'
-import Button from '../base/button'
+import Button from '../button'
 import OptionListItem from './common/option-list-item'
 import type { DatePickerProps } from './types'
 import { ViewType } from './types'
