+++ web/app/components/base/date-and-time-pickers/year-and-month-picker/footer.tsx
@@ -1,6 +1,6 @@
 import type { FC } from 'react'
 import React from 'react'
-import Button from '../../base/button'
+import Button from '../../button'
 import type { YearAndMonthPickerFooterProps } from '../types'
 
 const Footer: FC<YearAndMonthPickerFooterProps> = ({
